nvmem: core: add NVMEM_SYSFS Kconfig
Many nvmem providers are not very keen on having default sysfs
nvmem entry, as most of the usecases for them are inside kernel
itself. And in some cases read/writes to some areas in nvmem are
restricted and trapped at secure monitor level, so accessing them
from userspace would result in board reboots.

This patch adds new NVMEM_SYSFS Kconfig to make binary sysfs entry
an optional one. This provision will give more flexibility to users.
This patch also moves existing sysfs code to a new file so that its
not compiled in when its not really required.

Tested-by: Gaurav Kohli <gkohli@codeaurora.org>
Git-commit: ae0c2d725512f32a0d1a25f0cf2f07616d33a72e
Git-repo: https://git.kernel.org/pub/scm/linux/kernel/git/gregkh/char-misc.git
[schikk@codeaurora.org: Resolved merge conflicts and fixed
compilation issues]
